refactor: restructured home/pages directory
