+++ mobile/lib/presentation/home/pages/camera/camera.dart
